+++ packages/crdt_lf_hive/lib/src/adapters/ids.dart
@@ -1,9 +1,6 @@
 /// hlc adapter id
 const kHybridLogicalClockAdapter = 101;
 
-/// operation id adapter id
-const kOperationIdAdapter = 102;
-
 /// change adapter id
 const kChangeAdapter = 103;